+++ .php-cs-fixer.php
@@ -12,6 +12,7 @@ $finder = PhpCsFixer\Finder::create()
         'tests/Stub/fixtures',
         'examples/advanced/event_listener/access_attributes_of_task.php', // There is a class in it, we don't want it to be renamed
         'examples/advanced/remote_import/import_class.php', // There is a class in it, we don't want it to be renamed
+        'tests/Stub/StubsGeneratorTest.php', // see https://github.com/PHP-CS-Fixer/PHP-CS-Fixer/issues/9415
     ])
 ;